refactor: 🎨 theme selector constants change
